p4i integration Integrate 254147 from 4.3 to main: don't break if no output path is specified
[git-p4: depot-paths = "//depot/qt/main/": change = 254148]

+++ tests/arthur/htmlgenerator/htmlgenerator.cpp
@@ -417,7 +417,7 @@ void HTMLGenerator::convertToHtml()
                     image.flags |= Reference;
                 }
 
-                if (image.file.startsWith(outputDirName))
+                if (!outputDirName.isEmpty() && image.file.startsWith(outputDirName))
                     image.file.remove(0, outputDirName.length() + 1); // + '/'
                 HTMLSuite *htmlSuite = suites[suite->name];
                 if (!htmlSuite) {
